fix: 인앱 상품 구매 프로세스 개발
- 상품 조회
- 상품 주문 등록
- 상품 주문 소비

+++ apps/monglife-mongs-app-user-api/src/main/java/com/monglife/mongs/app/user/store/dto/etc/GetProductDto.java
@@ -14,10 +14,10 @@ public class GetProductDto {
 
     private String productName;
 
-    private Integer price;
+    private Double price;
 
     @Builder
-    public GetProductDto(String productId, String productName, Integer price) {
+    public GetProductDto(String productId, String productName, Double price) {
         this.productId = productId;
         this.productName = productName;
         this.price = price;
